status page markup separated from fetcher component
