+++ web/ui/react-app/src/pages/Status.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
-import { Status } from '.';
+import { statusRenderer as Status } from './Status';
 import { Alert } from 'reactstrap';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import * as useFetch from '../hooks/useFetches';
@@ -9,25 +9,17 @@ import toJson from 'enzyme-to-json';
 describe('Status', () => {
   afterEach(() => jest.restoreAllMocks());
   it('should render spinner while waiting data', () => {
-    const wrapper = shallow(<Status />);
+    const wrapper = shallow(<Status data={[] as any} />);
     expect(wrapper.find(FontAwesomeIcon)).toHaveLength(1);
   });
   it('should render Alert on error', () => {
     (useFetch as any).default = jest.fn().mockImplementation(() => ({ error: new Error('foo') }));
-    const wrapper = shallow(<Status />);
+    const wrapper = shallow(<Status error={{ message: 'foo' } as Error} />);
     expect(wrapper.find(Alert)).toHaveLength(1);
-  });
-  it('should fetch proper API endpoints', () => {
-    const useFetchSpy = jest.spyOn(useFetch, 'default');
-    shallow(<Status pathPrefix="/path/prefix" />);
-    expect(useFetchSpy).toHaveBeenCalledWith([
-      '/path/prefix/api/v1/status/runtimeinfo',
-      '/path/prefix/api/v1/status/buildinfo',
-      '/path/prefix/api/v1/alertmanagers',
-    ]);
+    expect(wrapper.find(Alert).prop('children')).toEqual([<strong>Error:</strong>, ' Error fetching status: ', 'foo']);
   });
   describe('Snapshot testing', () => {
-    const response = [
+    const response: any = [
       {
         startTime: '2019-10-30T22:03:23.247913868+02:00',
         CWD: '/home/boyskila/Desktop/prometheus',
@@ -64,7 +56,7 @@ describe('Status', () => {
     ];
     it('should match table snapshot', () => {
       (useFetch as any).default = jest.fn().mockImplementation(() => ({ response }));
-      const wrapper = shallow(<Status />);
+      const wrapper = shallow(<Status data={response} />);
       expect(toJson(wrapper)).toMatchSnapshot();
       jest.restoreAllMocks();
     });
